backend/pyproject.toml
[project]
name = "ai-design-studio-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sse-starlette>=2.1.0",
  "anthropic>=0.40.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/design_team"]

[tool.uv]
package = true
